requirements/base.txt
asgiref==3.6.0
Django==3.2.16
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-js-asset==2.0.0
Pillow==9.4.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
transliterate==1.10.2
psycopg2==2.9.5
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
django-dump-load-utf8==0.0.4
gunicorn==20.1.0
tzdata==2023.3
django-taggit==4.0.0
django-filter==23.2
social-auth-app-django==5.2.0
social-auth-core==4.4.2
django-extensions==3.2.1
pyOpenSSL==23.2.0
werkzeug==2.3.4
redis==4.5.5
django-elasticsearch-dsl==7.3
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1